Versao para a criacao da tag                  Os dialogs do service estão executando corretamente na Asyntask

+++ app/src/main/java/com/br/minasfrango/model/PrecoID.java
@@ -11,7 +11,7 @@ public class PrecoID extends RealmObject implements Serializable {
 	  @PrimaryKey
 		private long id;
 		
-		private double idProduto;
+		private long idProduto;
 		
 		private String unidadeProduto;
